MoneyUpdate

+++ CreditApplication/Models/CreditRate.cs
@@ -3,6 +3,6 @@
     public class CreditRate : BaseEntity
     {
         public string Name { get; set; }
-        public float MonthPercent { get; set; }
+        public decimal MonthPercent { get; set; }
     }
 }
